Realign release subpackages and have identity variant packages require variant packages
When the fedora-release-identity-<variant> packages were added, all
the content was moved indiscriminately from the fedora-release-<variant>
packages to the new fedora-release-identity-<variant> packages. This
meant that if someone installed one environment over another, the
necessary presets and system configuration files for the environment
to functionally work were not installed.

Fortunately, most of the fedora-release-<variant> packages previously
only contained identity information, so this wasn't a huge problem.
But, some of them contained environment presets and other configuration,
so let's fix this by shuffling things around so that this makes sense again.

Fixes: 6be4dd974eb4eba7e8233360a6354701fd153179

Resolves: rhbz#2148648